remove static_cast

+++ src/problems/RadhydroShockCGS/test_radhydro_shock_cgs.cpp
@@ -113,7 +113,7 @@ AMRSimulation<ShockProblem>::setCustomBoundaryConditions(const amrex::IntVect &i
 							 amrex::GeometryData const &geom, const amrex::Real /*time*/, const amrex::BCRec *bcr, int /*bcomp*/,
 							 int /*orig_comp*/)
 {
-	if ((bcr->lo(0) != static_cast<int>(quokka::BCType::ext_dir)) && (bcr->hi(0) != static_cast<int>(quokka::BCType::ext_dir))) {
+	if ((bcr->lo(0) != quokka::BCType::ext_dir) && (bcr->hi(0) != quokka::BCType::ext_dir)) {
 		return;
 	}
